+++ src/Phuntime/Aws/Type/ApiGatewayV2ProxyEvent.php
@@ -3,13 +3,20 @@ declare(strict_types=1);
 
 namespace Phuntime\Aws\Type;
 
+use Phuntime\Core\Contract\EventInterface;
+
 /**
  * @see https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/aws-lambda/trigger/api-gateway-proxy.d.ts
  * @see https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api-develop-integrations-lambda.html
  * @license MIT
  */
-class ApiGatewayV2ProxyEvent extends ApiGatewayProxyEvent
+class ApiGatewayV2ProxyEvent implements EventInterface
 {
+
+    protected string $httpMethod;
+    protected string $httpVersion;
+    protected string $path;
+    protected string $domainName;
 
     /**
      * @psalm-pure
@@ -19,7 +26,7 @@ class ApiGatewayV2ProxyEvent extends ApiGatewayProxyEvent
     public static function fromArray(array $payload): self
     {
         $object = new self();
-        $object->httpVersion = $payload['requestContext']['http']['method'];
+        $object->httpMethod = $payload['requestContext']['http']['method'];
         $object->path = $payload['requestContext']['http']['path'];
         $object->domainName = $payload['requestContext']['domainName'];
 
@@ -27,9 +34,51 @@ class ApiGatewayV2ProxyEvent extends ApiGatewayProxyEvent
     }
 
 
-
+    /**
+     * @deprecated
+     * @return string
+     */
     public function getVersion(): string
     {
         return '2.0';
     }
+
+    public function isAsync(): bool
+    {
+        return false;
+    }
+
+    /**
+     * @return string
+     */
+    public function getHttpVersion(): string
+    {
+        return $this->httpVersion;
+    }
+
+    /**
+     * @return string
+     */
+    public function getPath(): string
+    {
+        return $this->path;
+    }
+
+    /**
+     * @return string
+     */
+    public function getDomainName(): string
+    {
+        return $this->domainName;
+    }
+
+    /**
+     * @return string
+     */
+    public function getHttpMethod(): string
+    {
+        return $this->httpMethod;
+    }
+
+
 }